requirements.txt
numpy
click>=8.0
pillow
scipy==1.7.1
requests==2.26.0
tqdm==4.62.2
matplotlib==3.8.1
imageio==2.9.0
imgui==1.3.0
glfw==2.2.0
pyopengl==3.1.5
imageio-ffmpeg==0.4.3
opencv-python==4.6.0.66
lpips==0.1.4
kornia==0.6.8
pyspng
psutil
mrcfile
tensorboard
torch_efficient_distloss
ninja==1.10.2